Allow environment-specific Terraform to read cost centre parameter
Add global parameters policy which lets the environment-specific
Terraform scripts to read shared parameters set by the backend
Terraform.

This will let us set a shared parameter (like cost centre) in the
backend Terraform and use it in environment-specific Terraform scripts.

We might eventually replace this with a private git repo containing
global varibles, but this solution is easier to set up so we'll try it
for now.